Formatted code for good understanding

+++ src/DSCEngine.sol
@@ -28,8 +28,7 @@ pragma solidity ^0.8.18;
 import {DSC} from "./DSC.sol";
 import {ReentrancyGuard} from "./ReentrancyGuard.sol";
 import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
-import {AggregatorV3Interface}  from "./AggregatorV3Interface.sol";
-
+import {AggregatorV3Interface} from "./AggregatorV3Interface.sol";
 
 /*
 * @title: DSC
@@ -47,67 +46,70 @@ import {AggregatorV3Interface}  from "./AggregatorV3Interface.sol";
 * @notice This contract is very loosely based on the MakerDAO DSS (DAI) System.
 */
 
-contract DSCEngine is ReentrancyGuard{
-
-    /**ERRORS**/
+contract DSCEngine is ReentrancyGuard {
+    /**
+     * ERRORS*
+     */
     error DSCEngine__NeedsMoreThanZero();
     error DSCEngine__TokenAddressesAndPriceFeedAddressesMustBeSameLength();
     error DSCEngine__NotAllowedToken();
     error DSCEngine__TransferFailed();
 
-    /**State Variables**/
-
+    /**
+     * State Variables*
+     */
     uint256 private constant ADDITIONAL_FEED_PRECISION = 1e10;
     uint256 private constant PRECISION = 1e18;
 
-    mapping(address token =>address priceFeed) private s_priceFeeds;
+    mapping(address token => address priceFeed) private s_priceFeeds;
     mapping(address user => mapping(address token => uint256 amount)) s_collateralDeposited;
     mapping(address user => uint256 amountDscMinted) s_DSCMinted;
     address[] private s_collateralTokens;
 
     DSC private immutable i_dsc;
 
-
-    /**Events**/
-
+    /**
+     * Events*
+     */
     event CollateralDeposited(address indexed user, address indexed token, uint256 indexed amount);
 
-
-    /**Modifier**/
-    modifier moreThanZero(uint256 amount){
-        if(amount==0){
+    /**
+     * Modifier*
+     */
+    modifier moreThanZero(uint256 amount) {
+        if (amount == 0) {
             revert DSCEngine__NeedsMoreThanZero();
         }
         _;
     }
 
-    modifier isAllowedToken(address token){
-        if(s_priceFeeds[token]==address(0)){
+    modifier isAllowedToken(address token) {
+        if (s_priceFeeds[token] == address(0)) {
             revert DSCEngine__NotAllowedToken();
         }
         _;
     }
 
-    /**Constructor**/
-    constructor(
-        address[] memory tokenAddresses, 
-        address[] memory priceFeedAddresses, 
-        address dscAddress
-    ) {
+    /**
+     * Constructor*
+     */
+    constructor(address[] memory tokenAddresses, address[] memory priceFeedAddresses, address dscAddress) {
         //USD Price Feeds
-        if(tokenAddresses.length != priceFeedAddresses.length){
+        if (tokenAddresses.length != priceFeedAddresses.length) {
             revert DSCEngine__TokenAddressesAndPriceFeedAddressesMustBeSameLength();
         }
 
-        for(uint256 i = 0; i<tokenAddresses.length; i++){
+        for (uint256 i = 0; i < tokenAddresses.length; i++) {
             s_priceFeeds[tokenAddresses[i]] = priceFeedAddresses[i];
-            s_collateralTokens.push(tokenAddresses[i]); 
+            s_collateralTokens.push(tokenAddresses[i]);
         }
         i_dsc = DSC(dscAddress);
     }
 
-    /**Functions**/
-    function depositCollateralAndMintDsc() external{}
+    /**
+     * Functions*
+     */
+    function depositCollateralAndMintDsc() external {}
 
     /*
     *@notice follows CEI
@@ -115,43 +117,46 @@ contract DSCEngine is ReentrancyGuard{
     *@param amountCollateral The amount of collateral to deposit
     */
 
-    function depositCollateral(
-        address tokenCollateralAddress, 
-        uint256 amountCollateral
-    ) external moreThanZero(amountCollateral) isAllowedToken(tokenCollateralAddress) nonReentrant{
+    function depositCollateral(address tokenCollateralAddress, uint256 amountCollateral)
+        external
+        moreThanZero(amountCollateral)
+        isAllowedToken(tokenCollateralAddress)
+        nonReentrant
+    {
         s_collateralDeposited[msg.sender][tokenCollateralAddress] += amountCollateral;
         emit CollateralDeposited(msg.sender, tokenCollateralAddress, amountCollateral);
-        bool success  = IERC20(tokenCollateralAddress).transferFrom(msg.sender, address(this), amountCollateral);
-        if(!success){
+        bool success = IERC20(tokenCollateralAddress).transferFrom(msg.sender, address(this), amountCollateral);
+        if (!success) {
             revert DSCEngine__TransferFailed();
         }
     }
 
-    function redeemCollateral() external{}
+    function redeemCollateral() external {}
 
     /*
     *@notice follows CEI
     *@param amountDscToMint : The amount of DSC to mint
     *@notice they must have more collateral vaue than the threshold
     */
-    function mintDsc(uint256 amountDscToMint) external moreThanZero(amountDscToMint) nonReentrant{
-        s_DSCMinted[msg.sender]+=amountDscToMint;
+    function mintDsc(uint256 amountDscToMint) external moreThanZero(amountDscToMint) nonReentrant {
+        s_DSCMinted[msg.sender] += amountDscToMint;
     }
 
-    function redeemCollateralForDsc() external{}
+    function redeemCollateralForDsc() external {}
 
-    function burnDsc() external{}
+    function burnDsc() external {}
 
-    function liquidate() external{}
+    function liquidate() external {}
 
-    function getHealthFactor() external view{}
+    function getHealthFactor() external view {}
 
-    /**Private and Internal View Functions**/
-
-    function _getAccountInformation(address user) 
-        private 
-        view 
-        returns(uint256 totalDscMinted, uint256 collateralValueInUsd) 
+    /**
+     * Private and Internal View Functions*
+     */
+    function _getAccountInformation(address user)
+        private
+        view
+        returns (uint256 totalDscMinted, uint256 collateralValueInUsd)
     {
         totalDscMinted = s_DSCMinted[user];
         collateralValueInUsd = getAccountCollateralValue(user);
@@ -160,24 +165,24 @@ contract DSCEngine is ReentrancyGuard{
     //Returns how close to liquidation a user is
     //If a user gets below 1, then they can get liquidated
 
-    function _healthFactor(address user) private view returns (uint256){
+    function _healthFactor(address user) private view returns (uint256) {
         //Total DSC Minted
         //total collateral value
         (uint256 totalDscMinted, uint256 collateralValueInUsd) = _getAccountInformation(user);
-    }   
+    }
 
-    function _revertIfHealthFactorIsBrokens(address user) internal view{
+    function _revertIfHealthFactorIsBrokens(address user) internal view {
         // 1. Check health factor
         // 2. If health factor is broken, revert
     }
 
-
-    /** Public and External View Functions **/
-    
-    function getAccountCollateralValue(address user) public view returns(uint256){
+    /**
+     * Public and External View Functions *
+     */
+    function getAccountCollateralValue(address user) public view returns (uint256) {
         // loop through each collateral token, get the amount they have deposited, and map it to the price, to get the USD value
-        uint totalCollateralValueInUsd = 0;
-        for(uint256 i = 0; i < s_collateralTokens.length; i++){
+        uint256 totalCollateralValueInUsd = 0;
+        for (uint256 i = 0; i < s_collateralTokens.length; i++) {
             address token = s_collateralTokens[i];
             uint256 amount = s_collateralDeposited[user][token];
             totalCollateralValueInUsd += getUsdValue(token, amount);
@@ -185,9 +190,9 @@ contract DSCEngine is ReentrancyGuard{
         return totalCollateralValueInUsd;
     }
 
-    function getUsdValue(address token, uint256 amount) public view returns(uint256){
+    function getUsdValue(address token, uint256 amount) public view returns (uint256) {
         AggregatorV3Interface priceFeed = AggregatorV3Interface(s_priceFeeds[token]);
-        (,int256 price,,,) = priceFeed.latestRoundData();
+        (, int256 price,,,) = priceFeed.latestRoundData();
         return ((uint256(price) * ADDITIONAL_FEED_PRECISION) * amount) / PRECISION;
     }
 }
